requirements.txt
keras
keras_cv
tensorflow
tensorflow-datasets
Pillow
pycocotools
